EXPOSED-21 Primary key constraint not created (#1534)
Adjust equals() method in EntityIDColumnType.

IdTable creates a column of EntityIDColumnType to replace the original when
.entityId() is invoked. PrimaryKey() accepts arguments that can either be
the former or the original column of non-EntityIDColumnType. This results
in a failure to adequately compare column types when resolving a column's
description of its primary key constraint.

Both isOneColumnPK() and isLastColumnInPK() are also adjusted so that equals()
is invoked by the EntityIDColumnType (if present). This covers the
potential case when an IdTable may be created with a single primary key
column that has a user-defined name.