feat: EXPOSED-68 Add more json/json(b) column functions (#1770)
* feat: EXPOSED-68 Add more json/json(b) column functions

Add jsonContains() for PostgreSQL/NG and MySQL/MariaDB.

Add jsonExists() for all dialects except H2 and SQLServer. Note that a relevant
function does exist for SQLServer but it is only available as of 2022 v16.x,
while Exposed tests and supports 2017 v14.x

Add more unit tests.

Fix issues with jsonExtract():
- [Postgres] There are 2 functions, for json vs jsonb type, so a way to
distinguish between the invoking columns needed to be supported.
- Some dialects were hard-coded with the prefix '$.', but the root context item
alone '$' is a valid json path, so the period was removed and tests adjusted.
- Refactored dialect overrides to account for edge cases like: if no path args
provided; if too many path args provided; if optional arg provided when not
supported.

* feat: EXPOSED-68 Add more json/json(b) column functions

Add a jsonContains() override that allows a raw value to be provided that will
be wrapped to a literal under the hood.

Change how PostgreSQL checks whether a cast to JSONB needs to be made.