fix: EXPOSED-99 addMissingColumnsStatements comparison of datetime defaults (#1802)
The following test fails in exposed-javatime and exposed-kotlin-datetime modules 
(Note it was missing from exposed-jodatime module and has now been included):

DefaultTests/testConsistentSchemeWithFunctionAsDefaultExpression()

The test fails because an existing and unchanged datetime column with a database default is 
marked as modified, triggering an unnecessary ALTER statement. This is because the metadata 
value returned in addMissingColumnsStatements() does not match the returned value of dbDefaultToString():

- SQL Server maps both CURRENT_TIMESTAMP and GETDATE() to the ANSI standard equivalent, getdate.
- MariaDB returns CURRENT_TIMESTAMP(6) correctly, but in lower case.

Changes to the parsing of the default value cannot be done in DataTypeProvider.processForDefaultValue() 
because this function is also used in create statements, for example. So the processing needed to ensure 
a good match for these edge cases is confined to the existing private extension function in SchemaUtils -> DataTypeProvider.dbDefaultToString().