fix: Exception when using RESTRICT reference option (#1862)
* fix: Exception when using RESTRICT reference option with SQLServer

The following exception is thrown because SQLServer does not support RESTRICT reference option:
com.microsoft.sqlserver.jdbc.SQLServerException: Incorrect syntax near the keyword 'RESTRICT'.

* fix: Exception when using RESTRICT reference option with Oracle

The following exception is thrown because Oracle does not have the RESTRICT reference option:
java.sql.SQLSyntaxErrorException: ORA-00905: missing keyword

Preventing the deletion of a parent row if a child row references it is the default behaviour.

* fix: Allow ON UPDATE clause for H2 Oracle

All tests pass for it, so it seems like the ON UPDATE clause does work for H2 Oracle, but not for Oracle.

* chore: Change dialect names in DialectNameProvider to be more display-friendly

* chore: Simplify foreignKeyPart

* chore: Reuse testing logic in ForeignKeyConstraintTests