chore: EXPOSED-186 Replace JDK 1.7 support in exposed-jodatime classes (#1866)
* chore: Replace JDK 1.7 support in exposed-jodatime classes

Both column type classes in exposed-jodatime contain companion objects that check
for java.time classes that were only introduced since JDK 1.8. This was implemented
because the module was expected to run on Java version 1.7 and to be compatible
with legacy android versions.

Since the jvmToolchain has been to set to 8, this logic is no longer needed.

* chore: Replace JDK 1.7 support in exposed-jodatime classes

Add apiDump.