fix: EXPOSED-227 Slice() with empty list creates invalid SQL (#1899)
Providing an empty list to slice() resulted in a SELECT statement without any
columns defined, which returned syntax exceptions in all databases except
PostgreSQL and H2 (these returned empty results).

Adding require() to the query prepareSQL() allows an early fail in this
situation. The check could also go directly in an init block in the Slice
class. It could not go directly in the function itself, in the edge case
that a user uses Slice directly.