perf: EXPOSED-204 Performance problem with getConnection() (#1943)
* perf: EXPOSED-204 Performance problem with getConnection()

In spite of reused connections from a connection pool instance, the lazy
property initializer of a new database connection attempts to reset the
transaction isolation level. This occurs regardless of whether the connection
pool's set level has been overriden by a new value setting in DatabaseConfig
or on the transaction itself.

A flag in Database is now set if a DataSource is used to setup connections.
On the first connection, this flag ensures that the level set by the DataSource
is retrieved and cached. This avoids redundant resets in future connections
unless the DataSource value has been overriden by DatabaseConfig or per-transaction
values.

* perf: EXPOSED-204 Performance problem with getConnection()

Rename new properties.

Replace const integer value with a private flag that is passed from
ThreadLocalTransactionManager to ThreadLocalTransaction instance.

Edit tests to also check stored value in TransactionManager, in the event
users access this property directly, before the data source value is retrieved
and cached.

* perf: EXPOSED-204 Performance problem with getConnection()

Switch unitialized Database.dataSourceIsolationLevel from nullable to -1.