fix: EXPOSED-280 Comparison operators show incorrect compiler warning with datetime columns (#1984)
An incorrect compiler warning is flagged if a comparison operator, like eq(),
is used with 2 columns that store certain types. The warning treats the right-side
column as an EntityID column. The pattern so far seems to be stored types that are
not part of the Kotlin standard library, namely javatime, jodatime, and UUID types.

This occurs because of recently introduced convenience overrides that are being
inappropriately chosen. These overrides incorrectly make the ExpressionWithColumnType
argument's type parameter contravariant, when this annotation should instead be
set on the calling Expression object's type parameter.