feat: EXPOSED-238 Support EXPLAIN statements (#2022)
Add ExplainQuery class that processes EXPLAIN statements for all valid wrapped
statements. The constructor uses 2 new Transaction properties to ensure that the
wrapped statements are not executed and that the internal statements is passed to
the invoking instance.

Add an ExplainResultRow class so that the explain query can be executed and iterated
over in a similar manner to a standard query. This class currently is only good
for printing a readable string version of the results and needs getters.

Add unit tests.