fix: EXPOSED-411 ClassCastException when `uuid().references()` is used with `referrersOn` (#2127)
This fix allows using `referrersOn` or `backReferencedOn` in the parent table when a column of type `T` in the child table references a column of type `EntityID<T>` in the parent table.

This use case was not working even before any refactor was done. The location from which the error is coming _is_ different now after the [type safety refactor](https://github.com/JetBrains/Exposed/commit/42ac41b14c632e745c58d234641e9a2862d3204a), but it's essentially the same problem: the type mismatch.