feat: EXPOSED-459 Open AbstractQuery.copyTo() to allow custom Query class extension (#2173)
While it is possible to customize extension functions on the Query class, class
properties with private setters will not be copied to the new query instance when
chaining. So these custom functions could not be used with original queries.

This PR changes the visibility modifier of AbstractQuery.copyTo and overrides it
in the Query class.