feat!: EXPOSED-514 Support DELETE from tables in join (#2223)
* feat!: EXPOSED-514 Support DELETE from tables in join

UPDATE of a table from a join relation is already supported.

This adds the same functionality, but for DELETE. Rather than providing the values
to update, the specific target table to delete rows from must be specified.

Rather than create a new class, DeleteStatement has been refactored to take a
ColumnSet that could either be a Table or a Join (like UpdateStatement). So the
table property and old constructor have been deprecated in favor of a new property
targetsSet and a constructor with the new property and a targetTables property.