fix!: EXPOSED-536 Short column allows out-of-range values (#2231)
For SQLite and Oracle, inserting an out-of-range value using raw SQL with `exec` would work even though it shouldn't.

Short stores values between -32768 and 32767. For Oracle and H2 Oracle, Short data type was changed from SMALLINT to NUMBER(5) because SMALLINT is stored as NUMBER(38) in the database and takes up unnecessary storage. Another reason for this change is that this is a precursor to the column type change detection feature for migration, where it will be useful to have the column type be more specific than NUMBER(38).

A check constraint was added (that applies only to SQLite, Oracle and H2_Oracle) to ensure that no out-of-range values are stored.