fix!: EXPOSED-545 Byte column allows out-of-range values
For SQLite, SQL Server, Oracle, PostgreSQL, and H2 PostgreSQL, inserting an out-of-range value using raw SQL with exec would work even though it shouldn't.

Byte stores values between -128 and 127. For Oracle, Byte data type was changed from SMALLINT to NUMBER(3) because SMALLINT is stored as NUMBER(38) in the database and takes up unnecessary storage. Another reason for this change is that this is a precursor to the column type change detection feature for migration, where it will be useful to have the column type be more specific than NUMBER(38).

A check constraint was added (that applies only to SQL Server, SQLite, Oracle, PostgreSQL, and H2 PostgreSQL) to ensure that no out-of-range values are stored.

For SQL Server, Byte data type was changed from TINYINT to SMALLINT. Using TINYINT is incorrect because it [allows values from 0 to 255](https://learn.microsoft.com/en-us/sql/t-sql/data-types/int-bigint-smallint-and-tinyint-transact-sql?view=sql-server-ver16#:~:text=2%20bytes-,tinyint,-0%20to%20255). An additional change was needed for SQL Server because when attempting to insert any value below 0 using Exposed, it fails because of the CHECK constraint even though it shouldn't. When inserting using exec, however, it works as expected. Updating the SQL Server database version in the docker compose file and the SQL Server JDBC driver version did not fix it. The problem is most likely in the way that the driver handles the Byte type, so as a workaround, overriding `valueToDB` in `ByteColumnType` and converting the Byte value to Short solves the issue.