fix: EXPOSED-621 IllegalStateException on accessing autoincrement column after insert using Entity (#2291)
The issue was that the `hashCode()` function was using `sequence` which invokes `autoincSeq` if `_sequence` is null, which then invokes `currentDialect`, and if this is not within a transaction, it will throw an error. This was introduced in the pull request for EXPOSED-487. The fix is to use `_sequence` instead of `sequence` when generating a hashcode.