feat: EXPOSED-733 Detect column type change for migrations in H2 (#2419)
What:
Column type changes in H2 are now detected when generating database migration statements.

How:
ColumnMetadata now has a new property, sqlType. To get this type, the column types pre-fetched (if the database supports that) in the extractColumns function in JdbcDatabaseMetadataImpl. If pre-fetching is not supported, the new getColumnType function queries the database for information about the column (future PRs).
Some SQL types are mapped to a different type in the database than one might expect, so normalizing the type name is done in those cases. Let me know if you think the normalizedColumnType function in H2 should be moved to DatabaseDialect and turned into an open function that users can override.
When checking for column differences, the type of the existing column is compared with the type of the Exposed table's column using the new function areEquivalentColumnTypes.